Huge rename applovin to maxsdk
