+++ extension-maxsdk/ext.manifest
