+++ extension-maxsdk/manifests/android/AndroidManifest.xml
