+++ extension-maxsdk/manifests/android/build.gradle
