+++ extension-maxsdk/src/android/com_defold_maxsdk_AppLovinMaxJNI.h
@@ -1,5 +1,5 @@
 #include <jni.h>
-/* Header for class com_defold_applovin_AppLovinMaxJNI */
+/* Header for class com_defold_maxsdk_AppLovinMaxJNI */
 
 #ifndef COM_DEFOLD_APPLOVIN_APPLOVINMAXJNI_H
 #define COM_DEFOLD_APPLOVIN_APPLOVINMAXJNI_H
@@ -7,11 +7,11 @@
 extern "C" {
 #endif
 	/*
-	* Class:     com_defold_applovin_AppLovinMaxJNI
-	* Method:    applovinAddToQueue_first_arg
+	* Class:     com_defold_maxsdk_AppLovinMaxJNI
+	* Method:    maxsdkAddToQueue_first_arg
 	* Signature: (ILjava/lang/String;I)V
 	*/
-	JNIEXPORT void JNICALL Java_com_defold_applovin_AppLovinMaxJNI_applovinAddToQueue
+	JNIEXPORT void JNICALL Java_com_defold_maxsdk_AppLovinMaxJNI_maxsdkAddToQueue
 		(JNIEnv *, jclass, jint, jstring);
 
 #ifdef __cplusplus
