+++ extension-maxsdk/src/android/maxsdk_jni.h
