+++ extension-maxsdk/src/maxsdk_private.h
