+++ game.project
@@ -1,5 +1,5 @@
 [bootstrap]
-main_collection = /main/applovin_example.collectionc
+main_collection = /main/maxsdk_example.collectionc
 
 [input]
 game_binding = /input/game.input_bindingc
@@ -17,7 +17,7 @@ title = AppLovinMax
 dependencies = https://github.com/andsve/dirtylarry/archive/master.zip,https://github.com/defold/extension-adinfo/archive/refs/tags/3.0.0.zip
 
 [library]
-include_dirs = applovin
+include_dirs = extension-maxsdk
 
 [applovin]
 sdk_key_android = OquH2hNFtGlwEoR03PJa3xdoXRnuHt5Lstpk0yMREvigNeVdmmR6BQdwx_6x8LazqQ_tHMa9RN4cNVkrg2b8Lm
